fix(core): 修复内存泄漏、截断、空指针守卫与返回值检查
- McpToolRegistry: _exit_tree() 补 memdelete（单例泄漏）
- game_bridge: int64_t id 消除 32 位截断
- tool_adapter: 非 Dictionary 返回值走错误路径而非静默丢弃
- create_resource: ResourceSaver::save() 返回值检查（唯一遗漏）
- get_performance_monitors: EditorInterface 空守卫
- mcp_console/mcp_dock: DisplayServer 空守卫
- add_animation_node: position schema 补 x/y 子属性结构指引

+++ extensions/src/runtime/game_bridge.cpp
@@ -202,7 +202,9 @@ void GameBridgeNode::read_clients() {
     String cmd = msg_dict.get("cmd", "");
     Dictionary params = msg_dict.get("params", Dictionary());
     Variant id_v = msg_dict.get("id", Variant());
-    int id = static_cast<int>((id_v.get_type() == Variant::FLOAT ? static_cast<double>(id_v ): static_cast<int64_t>(id_v)));
+    int64_t id = (id_v.get_type() == Variant::FLOAT
+        ? static_cast<int64_t>(static_cast<double>(id_v))
+        : static_cast<int64_t>(id_v));
 
     Dictionary result = dispatch(cmd, params);
     result["id"] = id;
